requirement.txt
pytesseract
pillow
ttkbootstrap
tkinterdnd2
